# 🚀 SIMD optimization: AVX2 vectorization with major performance improvements