--- a/src/lib.rs
+++ b/src/lib.rs
@@ -18,16 +18,20 @@ pub use implementations::{
     blocked_matmul_default,
     blocked_matmul_with_dotprod,
     blocked_matmul_col_major_fast,
-    blocked_matmul_optimized
+    blocked_matmul_optimized,
+    simd_matmul,
+    simd_blocked_matmul_optimized,
+    simd_blocked_matmul_32,
+    simd_blocked_matmul_128
 };
-pub use dotprod::{naive_dotprod, unrolled_dotprod};
+pub use dotprod::{naive_dotprod, unrolled_dotprod, simd_dotprod};
 
 #[cfg(test)]
 mod benches {
     extern crate test;
     use test::Bencher;
     use super::*;
-    use nalgebra::{DMatrix};
+    use nalgebra::{DMatrix, DVector};
 
     // Helper function to convert our Matrix to nalgebra DMatrix
     fn matrix_to_dmatrix(mat: &Matrix) -> DMatrix<f64> {
@@ -38,6 +42,11 @@ mod benches {
             }
         }
         DMatrix::from_row_slice(mat.rows, mat.cols, &data)
+    }
+
+    // Helper function to convert Vec<f64> to nalgebra DVector
+    fn vec_to_dvector(vec: &[f64]) -> DVector<f64> {
+        DVector::from_vec(vec.to_vec())
     }
 
     #[bench]
@@ -198,7 +207,7 @@ mod benches {
     }
 
     #[bench]
-    fn bench_dotprod_matmul_naive_64x64(b: &mut Bencher) {
+    fn bench_mm_with_naive_dotprod_64x64(b: &mut Bencher) {
         let a = Matrix::random(64, 64);
         let mat_b = Matrix::random(64, 64);
         b.iter(|| {
@@ -207,7 +216,7 @@ mod benches {
     }
 
     #[bench]
-    fn bench_dotprod_matmul_unrolled_64x64(b: &mut Bencher) {
+    fn bench_mm_with_unrolled_dotprod_64x64(b: &mut Bencher) {
         let a = Matrix::random(64, 64);
         let mat_b = Matrix::random(64, 64);
         b.iter(|| {
@@ -216,7 +225,7 @@ mod benches {
     }
 
     #[bench]
-    fn bench_dotprod_matmul_naive_128x128(b: &mut Bencher) {
+    fn bench_mm_with_naive_dotprod_128x128(b: &mut Bencher) {
         let a = Matrix::random(128, 128);
         let mat_b = Matrix::random(128, 128);
         b.iter(|| {
@@ -225,7 +234,7 @@ mod benches {
     }
 
     #[bench]
-    fn bench_dotprod_matmul_unrolled_128x128(b: &mut Bencher) {
+    fn bench_mm_with_unrolled_dotprod_128x128(b: &mut Bencher) {
         let a = Matrix::random(128, 128);
         let mat_b = Matrix::random(128, 128);
         b.iter(|| {
@@ -234,7 +243,7 @@ mod benches {
     }
 
     #[bench]
-    fn bench_dotprod_matmul_fast_naive_128x128(b: &mut Bencher) {
+    fn bench_mm_fast_with_naive_dotprod_128x128(b: &mut Bencher) {
         let a = Matrix::random(128, 128);
         let mat_b = Matrix::random(128, 128);
         b.iter(|| {
@@ -243,7 +252,7 @@ mod benches {
     }
 
     #[bench]
-    fn bench_dotprod_matmul_fast_unrolled_128x128(b: &mut Bencher) {
+    fn bench_mm_fast_with_unrolled_dotprod_128x128(b: &mut Bencher) {
         let a = Matrix::random(128, 128);
         let mat_b = Matrix::random(128, 128);
         b.iter(|| {
@@ -252,7 +261,7 @@ mod benches {
     }
 
     #[bench]
-    fn bench_dotprod_matmul_fast_naive_64x64(b: &mut Bencher) {
+    fn bench_mm_fast_with_naive_dotprod_64x64(b: &mut Bencher) {
         let a = Matrix::random(64, 64);
         let mat_b = Matrix::random(64, 64);
         b.iter(|| {
@@ -261,7 +270,7 @@ mod benches {
     }
 
     #[bench]
-    fn bench_dotprod_matmul_fast_unrolled_64x64(b: &mut Bencher) {
+    fn bench_mm_fast_with_unrolled_dotprod_64x64(b: &mut Bencher) {
         let a = Matrix::random(64, 64);
         let mat_b = Matrix::random(64, 64);
         b.iter(|| {
@@ -270,7 +279,7 @@ mod benches {
     }
 
     #[bench]
-    fn bench_dotprod_matmul_col_major_fast_unrolled_64x64(b: &mut Bencher) {
+    fn bench_mm_col_major_fast_with_unrolled_dotprod_64x64(b: &mut Bencher) {
         let a = Matrix::random(64, 64);
         let mat_b = Matrix::random(64, 64);
         b.iter(|| {
@@ -279,7 +288,7 @@ mod benches {
     }
 
     #[bench]
-    fn bench_dotprod_matmul_col_major_fast_unrolled_128x128(b: &mut Bencher) {
+    fn bench_mm_col_major_fast_with_unrolled_dotprod_128x128(b: &mut Bencher) {
         let a = Matrix::random(128, 128);
         let mat_b = Matrix::random(128, 128);
         b.iter(|| {
@@ -288,7 +297,7 @@ mod benches {
     }
 
     #[bench]
-    fn bench_dotprod_matmul_col_major_fast_naive_64x64(b: &mut Bencher) {
+    fn bench_mm_col_major_fast_with_naive_dotprod_64x64(b: &mut Bencher) {
         let a = Matrix::random(64, 64);
         let mat_b = Matrix::random(64, 64);
         b.iter(|| {
@@ -297,7 +306,7 @@ mod benches {
     }
 
     #[bench]
-    fn bench_dotprod_matmul_col_major_fast_naive_128x128(b: &mut Bencher) {
+    fn bench_mm_col_major_fast_with_naive_dotprod_128x128(b: &mut Bencher) {
         let a = Matrix::random(128, 128);
         let mat_b = Matrix::random(128, 128);
         b.iter(|| {
@@ -427,11 +436,181 @@ mod benches {
     }
 
     #[bench]
+    fn bench_simd_128x128(b: &mut Bencher) {
+        let a = Matrix::random(128, 128);
+        let mat_b = Matrix::random(128, 128);
+        b.iter(|| {
+            test::black_box(simd_matmul(&a, &mat_b))
+        });
+    }
+
+    #[bench]
+    fn bench_simd_256x256(b: &mut Bencher) {
+        let a = Matrix::random(256, 256);
+        let mat_b = Matrix::random(256, 256);
+        b.iter(|| {
+            test::black_box(simd_matmul(&a, &mat_b))
+        });
+    }
+
+    #[bench]
+    fn bench_simd_512x512(b: &mut Bencher) {
+        let a = Matrix::random(512, 512);
+        let mat_b = Matrix::random(512, 512);
+        b.iter(|| {
+            test::black_box(simd_matmul(&a, &mat_b))
+        });
+    }
+
+    #[bench]
+    fn bench_simd_blocked_optimized_256x256(b: &mut Bencher) {
+        let a = Matrix::random(256, 256);
+        let mat_b = Matrix::random(256, 256);
+        b.iter(|| {
+            test::black_box(simd_blocked_matmul_optimized(&a, &mat_b))
+        });
+    }
+
+    #[bench]
+    fn bench_simd_blocked_optimized_512x512(b: &mut Bencher) {
+        let a = Matrix::random(512, 512);
+        let mat_b = Matrix::random(512, 512);
+        b.iter(|| {
+            test::black_box(simd_blocked_matmul_optimized(&a, &mat_b))
+        });
+    }
+
+    #[bench]
+    fn bench_simd_blocked_32_512x512(b: &mut Bencher) {
+        let a = Matrix::random(512, 512);
+        let mat_b = Matrix::random(512, 512);
+        b.iter(|| {
+            test::black_box(simd_blocked_matmul_32(&a, &mat_b))
+        });
+    }
+
+    #[bench]
+    fn bench_simd_blocked_128_512x512(b: &mut Bencher) {
+        let a = Matrix::random(512, 512);
+        let mat_b = Matrix::random(512, 512);
+        b.iter(|| {
+            test::black_box(simd_blocked_matmul_128(&a, &mat_b))
+        });
+    }
+
+    #[bench]
     fn bench_blocked_optimized_512x512(b: &mut Bencher) {
         let a = Matrix::random(512, 512);
         let mat_b = Matrix::random(512, 512);
         b.iter(|| {
             test::black_box(blocked_matmul_optimized(&a, &mat_b, unrolled_dotprod))
+        });
+    }
+
+    // Dot Product Benchmarks
+    
+    #[bench]
+    fn bench_vector_dotprod_naive_128(b: &mut Bencher) {
+        let a = (0..128).map(|i| i as f64 * 0.1).collect::<Vec<_>>();
+        let vec_b = (0..128).map(|i| (i + 1) as f64 * 0.2).collect::<Vec<_>>();
+        b.iter(|| {
+            test::black_box(naive_dotprod(&a, &vec_b))
+        });
+    }
+
+    #[bench]
+    fn bench_vector_dotprod_unrolled_128(b: &mut Bencher) {
+        let a = (0..128).map(|i| i as f64 * 0.1).collect::<Vec<_>>();
+        let vec_b = (0..128).map(|i| (i + 1) as f64 * 0.2).collect::<Vec<_>>();
+        b.iter(|| {
+            test::black_box(unrolled_dotprod(&a, &vec_b))
+        });
+    }
+
+    #[bench]
+    fn bench_vector_dotprod_simd_128(b: &mut Bencher) {
+        let a = (0..128).map(|i| i as f64 * 0.1).collect::<Vec<_>>();
+        let vec_b = (0..128).map(|i| (i + 1) as f64 * 0.2).collect::<Vec<_>>();
+        b.iter(|| {
+            test::black_box(simd_dotprod(&a, &vec_b))
+        });
+    }
+
+    #[bench]
+    fn bench_vector_dotprod_nalgebra_128(b: &mut Bencher) {
+        let a = (0..128).map(|i| i as f64 * 0.1).collect::<Vec<_>>();
+        let vec_b = (0..128).map(|i| (i + 1) as f64 * 0.2).collect::<Vec<_>>();
+        let dvec_a = vec_to_dvector(&a);
+        let dvec_b = vec_to_dvector(&vec_b);
+        b.iter(|| {
+            test::black_box(dvec_a.dot(&dvec_b))
+        });
+    }
+
+    #[bench]
+    fn bench_vector_dotprod_naive_1024(b: &mut Bencher) {
+        let a = (0..1024).map(|i| i as f64 * 0.1).collect::<Vec<_>>();
+        let vec_b = (0..1024).map(|i| (i + 1) as f64 * 0.2).collect::<Vec<_>>();
+        b.iter(|| {
+            test::black_box(naive_dotprod(&a, &vec_b))
+        });
+    }
+
+    #[bench]
+    fn bench_vector_dotprod_unrolled_1024(b: &mut Bencher) {
+        let a = (0..1024).map(|i| i as f64 * 0.1).collect::<Vec<_>>();
+        let vec_b = (0..1024).map(|i| (i + 1) as f64 * 0.2).collect::<Vec<_>>();
+        b.iter(|| {
+            test::black_box(unrolled_dotprod(&a, &vec_b))
+        });
+    }
+
+    #[bench]
+    fn bench_vector_dotprod_simd_1024(b: &mut Bencher) {
+        let a = (0..1024).map(|i| i as f64 * 0.1).collect::<Vec<_>>();
+        let vec_b = (0..1024).map(|i| (i + 1) as f64 * 0.2).collect::<Vec<_>>();
+        b.iter(|| {
+            test::black_box(simd_dotprod(&a, &vec_b))
+        });
+    }
+
+    #[bench]
+    fn bench_vector_dotprod_nalgebra_1024(b: &mut Bencher) {
+        let a = (0..1024).map(|i| i as f64 * 0.1).collect::<Vec<_>>();
+        let vec_b = (0..1024).map(|i| (i + 1) as f64 * 0.2).collect::<Vec<_>>();
+        let dvec_a = vec_to_dvector(&a);
+        let dvec_b = vec_to_dvector(&vec_b);
+        b.iter(|| {
+            test::black_box(dvec_a.dot(&dvec_b))
+        });
+    }
+
+    #[bench]
+    fn bench_vector_dotprod_naive_4096(b: &mut Bencher) {
+        let a = (0..4096).map(|i| i as f64 * 0.1).collect::<Vec<_>>();
+        let vec_b = (0..4096).map(|i| (i + 1) as f64 * 0.2).collect::<Vec<_>>();
+        b.iter(|| {
+            test::black_box(naive_dotprod(&a, &vec_b))
+        });
+    }
+
+    #[bench]
+    fn bench_vector_dotprod_simd_4096(b: &mut Bencher) {
+        let a = (0..4096).map(|i| i as f64 * 0.1).collect::<Vec<_>>();
+        let vec_b = (0..4096).map(|i| (i + 1) as f64 * 0.2).collect::<Vec<_>>();
+        b.iter(|| {
+            test::black_box(simd_dotprod(&a, &vec_b))
+        });
+    }
+
+    #[bench]
+    fn bench_vector_dotprod_nalgebra_4096(b: &mut Bencher) {
+        let a = (0..4096).map(|i| i as f64 * 0.1).collect::<Vec<_>>();
+        let vec_b = (0..4096).map(|i| (i + 1) as f64 * 0.2).collect::<Vec<_>>();
+        let dvec_a = vec_to_dvector(&a);
+        let dvec_b = vec_to_dvector(&vec_b);
+        b.iter(|| {
+            test::black_box(dvec_a.dot(&dvec_b))
         });
     }
 }